+++ rest_auth/registration/views.py
@@ -5,22 +5,29 @@ from django.views.decorators.debug import sensitive_post_parameters
 
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework.permissions import AllowAny
-from rest_framework.generics import CreateAPIView
+from rest_framework.permissions import (AllowAny,
+                                        IsAuthenticated)
+from rest_framework.generics import CreateAPIView, ListAPIView, GenericAPIView
+from rest_framework.exceptions import NotFound
 from rest_framework import status
 
 from allauth.account.adapter import get_adapter
 from allauth.account.views import ConfirmEmailView
 from allauth.account.utils import complete_signup
 from allauth.account import app_settings as allauth_settings
+from allauth.socialaccount import signals
+from allauth.socialaccount.adapter import get_adapter as get_social_adapter
+from allauth.socialaccount.models import SocialAccount
 
 from rest_auth.app_settings import (UserDetailsSerializer,
                                     TokenSerializer,
                                     JWTSerializer,
                                     create_token)
 from rest_auth.models import TokenModel
-from rest_auth.registration.serializers import (SocialLoginSerializer,
-                                                VerifyEmailSerializer)
+from rest_auth.registration.serializers import (VerifyEmailSerializer,
+                                                SocialLoginSerializer,
+                                                SocialAccountSerializer,
+                                                SocialConnectSerializer)
 from rest_auth.utils import jwt_encode
 from rest_auth.views import LoginView
 from .app_settings import RegisterSerializer, register_permission_classes
@@ -113,12 +120,70 @@ class SocialLoginView(LoginView):
 
     class FacebookLogin(SocialLoginView):
         adapter_class = FacebookOAuth2Adapter
-         client_class = OAuth2Client
-         callback_url = 'localhost:8000'
+        client_class = OAuth2Client
+        callback_url = 'localhost:8000'
     -------------
     """
-
     serializer_class = SocialLoginSerializer
 
     def process_login(self):
         get_adapter(self.request).login(self.request, self.user)
+
+
+class SocialConnectView(LoginView):
+    """
+    class used for social account linking
+
+    example usage for facebook with access_token
+    -------------
+    from allauth.socialaccount.providers.facebook.views import FacebookOAuth2Adapter
+
+    class FacebookConnect(SocialConnectView):
+        adapter_class = FacebookOAuth2Adapter
+    -------------
+    """
+    serializer_class = SocialConnectSerializer
+    permission_classes = (IsAuthenticated,)
+
+    def process_login(self):
+        get_adapter(self.request).login(self.request, self.user)
+
+
+class SocialAccountListView(ListAPIView):
+    """
+    List SocialAccounts for the currently logged in user
+    """
+    serializer_class = SocialAccountSerializer
+    permission_classes = (IsAuthenticated,)
+
+    def get_queryset(self):
+        return SocialAccount.objects.filter(user=self.request.user)
+
+
+class SocialAccountDisconnectView(GenericAPIView):
+    """
+    Disconnect SocialAccount from remote service for
+    the currently logged in user
+    """
+    serializer_class = SocialConnectSerializer
+    permission_classes = (IsAuthenticated,)
+
+    def get_queryset(self):
+        return SocialAccount.objects.filter(user=self.request.user)
+
+    def post(self, request, *args, **kwargs):
+        accounts = self.get_queryset()
+        account = accounts.filter(pk=kwargs['pk']).first()
+        if not account:
+            raise NotFound
+
+        get_social_adapter(self.request).validate_disconnect(account, accounts)
+
+        account.delete()
+        signals.social_account_removed.send(
+            sender=SocialAccount,
+            request=self.request,
+            socialaccount=account
+        )
+
+        return Response(self.get_serializer(account).data)