+++ rest_auth/tests/mixins.py
@@ -1,11 +1,23 @@
 import json
 
 from django.conf import settings
-from django.core.urlresolvers import reverse
 from django.test.client import Client, MULTIPART_CONTENT
 from django.utils.encoding import force_text
 
 from rest_framework import status
+from rest_framework import permissions
+
+try:
+    from django.urls import reverse
+except ImportError:
+    from django.core.urlresolvers import reverse
+
+
+class CustomPermissionClass(permissions.BasePermission):
+    message = 'You shall not pass!'
+
+    def has_permission(self, request, view):
+        return False
 
 
 class APIClient(Client):
@@ -17,8 +29,7 @@ class APIClient(Client):
         return self.generic('OPTIONS', path, data, content_type, **extra)
 
 
-class BaseAPITestCase(object):
-
+class TestsMixin(object):
     """
     base for API tests:
         * easy request calls, f.e.: self.post(url, data), self.get(url)
@@ -64,29 +75,6 @@ class BaseAPITestCase(object):
     def patch(self, *args, **kwargs):
         return self.send_request('patch', *args, **kwargs)
 
-    # def put(self, *args, **kwargs):
-    #     return self.send_request('put', *args, **kwargs)
-
-    # def delete(self, *args, **kwargs):
-    #     return self.send_request('delete', *args, **kwargs)
-
-    # def options(self, *args, **kwargs):
-    #     return self.send_request('options', *args, **kwargs)
-
-    # def post_file(self, *args, **kwargs):
-    #     kwargs['content_type'] = MULTIPART_CONTENT
-    #     return self.send_request('post', *args, **kwargs)
-
-    # def get_file(self, *args, **kwargs):
-    #     content_type = None
-    #     if 'content_type' in kwargs:
-    #         content_type = kwargs.pop('content_type')
-    #     response = self.send_request('get', *args, **kwargs)
-    #     if content_type:
-    #         self.assertEqual(
-    #             bool(filter(lambda x: content_type in x, response._headers['content-type'])), True)
-    #     return response
-
     def init(self):
         settings.DEBUG = True
         self.client = APIClient()
@@ -102,6 +90,9 @@ class BaseAPITestCase(object):
         self.tw_login_url = reverse('tw_login')
         self.tw_login_no_view_url = reverse('tw_login_no_view')
         self.tw_login_no_adapter_url = reverse('tw_login_no_adapter')
+        self.fb_connect_url = reverse('fb_connect')
+        self.tw_connect_url = reverse('tw_connect')
+        self.social_account_list_url = reverse('social_account_list')
 
     def _login(self):
         payload = {