+++ rest_auth/views.py
@@ -89,11 +89,22 @@ class LoginView(GenericAPIView):
             serializer = serializer_class(instance=self.user,
                                           context={'request': self.request})
 
-        return Response(serializer.data, status=status.HTTP_200_OK)
+        response = Response(serializer.data, status=status.HTTP_200_OK)
+        if getattr(settings, 'REST_USE_JWT', False):
+            from rest_framework_jwt.settings import api_settings as jwt_settings
+            if jwt_settings.JWT_AUTH_COOKIE:
+                from datetime import datetime
+                expiration = (datetime.utcnow() + jwt_settings.JWT_EXPIRATION_DELTA)
+                response.set_cookie(jwt_settings.JWT_AUTH_COOKIE,
+                                    self.token,
+                                    expires=expiration,
+                                    httponly=True)
+        return response
 
     def post(self, request, *args, **kwargs):
         self.request = request
-        self.serializer = self.get_serializer(data=self.request.data)
+        self.serializer = self.get_serializer(data=self.request.data,
+                                              context={'request': request})
         self.serializer.is_valid(raise_exception=True)
 
         self.login()
@@ -117,7 +128,7 @@ class LogoutView(APIView):
 
         return self.finalize_response(request, response, *args, **kwargs)
 
-    def post(self, request):
+    def post(self, request, *args, **kwargs):
         return self.logout(request)
 
     def logout(self, request):
@@ -125,11 +136,16 @@ class LogoutView(APIView):
             request.user.auth_token.delete()
         except (AttributeError, ObjectDoesNotExist):
             pass
+        if getattr(settings, 'REST_SESSION_LOGIN', True):
+            django_logout(request)
 
-        django_logout(request)
-
-        return Response({"detail": _("Successfully logged out.")},
-                        status=status.HTTP_200_OK)
+        response = Response({"detail": _("Successfully logged out.")},
+                            status=status.HTTP_200_OK)
+        if getattr(settings, 'REST_USE_JWT', False):
+            from rest_framework_jwt.settings import api_settings as jwt_settings
+            if jwt_settings.JWT_AUTH_COOKIE:
+                response.delete_cookie(jwt_settings.JWT_AUTH_COOKIE)
+        return response
 
 
 class UserDetailsView(RetrieveUpdateAPIView):
@@ -197,7 +213,7 @@ class PasswordResetConfirmView(GenericAPIView):
     def dispatch(self, *args, **kwargs):
         return super(PasswordResetConfirmView, self).dispatch(*args, **kwargs)
 
-    def post(self, request):
+    def post(self, request, *args, **kwargs):
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
@@ -220,7 +236,7 @@ class PasswordChangeView(GenericAPIView):
     def dispatch(self, *args, **kwargs):
         return super(PasswordChangeView, self).dispatch(*args, **kwargs)
 
-    def post(self, request):
+    def post(self, request, *args, **kwargs):
         serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()